requirements.txt
python-dotenv==0.19.2
openai>=1.3.0
tiktoken
# numpy>=1.24.0 install manually first before running this file
pandas>=1.5.0
pytest==7.0.0 